chore(entry): convert main/index to TSX (add .bak, stubs)
