+++ frontend/src/index.jsx
@@ -1,10 +1,1 @@
-// Minimal frontend entrypoint used during typecheck
-import React from 'react';
-import { createRoot } from 'react-dom/client';
-import App from './App';
-
-const container = document.getElementById('root');
-if (container) {
-	const root = createRoot(container);
-	root.render(<App />);
-}
+export { } from './index.tsx';
