+++ frontend/src/index.jsx.bak
@@ -1,0 +1,11 @@
+// Backup of index.jsx before conversion to TypeScript
+// Minimal frontend entrypoint used during typecheck
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import App from './App';
+
+const container = document.getElementById('root');
+if (container) {
+    const root = createRoot(container);
+    root.render(<App />);
+}
