+++ frontend/src/main.jsx.bak
@@ -1,0 +1,11 @@
+// Backup of main.jsx before conversion to TypeScript
+import React from "react";
+import ReactDOM from "react-dom/client";
+import App from "./App";
+import "./index.css";
+
+ReactDOM.createRoot(document.getElementById("root")).render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
